remove vendor/ from source control
