+++ package.json
@@ -1,7 +1,8 @@
 {
   "name": "@enhance/create",
   "version": "2.3.0",
-  "main": "index.mjs",
+  "main": "index.js",
+  "type": "module",
   "scripts": {
     "vendor": "./scripts/vendor.sh"
   },
